### 2024-05-25 #644 @serkor1
Presubmission Inquiry of the `cryptoQuotes`-package
Submitting Author Name: Serkan Korkmaz
Submitting Author Github Handle: @serkor1
Repository:  https://github.com/serkor1/cryptoQuotes
Submission type: Pre-submission

---

```
Package: cryptoQuotes
Title: A Streamlined Access to Cryptocurrency OHLC-V Market Data and Sentiment Indicators
Version: 1.3.1
    person("Serkan", "Korkmaz", , "serkor1@duck.com", role = c("cre", "aut", "ctb", "cph"),
    person("Jonas", "Cuzulan Hirani", , "jjh@vive.dk", role = "ctb",
  )
  This high-level API client offers a streamlined access to public cryptocurrency market data and sentiment indicators. It features OHLC-V (Open, High, Low, Close, Volume) that comes
  with granularity ranging from seconds to months and essential sentiment indicators to develop and backtest trading strategies, or conduct detailed market analysis. By interacting directly with
  the major cryptocurrency exchanges this package ensures a reliable, and stable, flow of market information, eliminating the need for complex, low-level API interactions or webcrawlers.
    quantmod,
    tidyverse
    cli (>= 3.6.2),
    curl (>= 5.2.1),
    jsonlite (>= 1.8.8),
    lifecycle (>= 1.0.4),
    plotly (>= 4.10.4),
    TTR (>= 0.24.4),
    xts (>= 0.13.2),
    zoo (>= 1.8-12)

```

The package retrieves cryptocurrency market data by interacting with public API endpoints provided by the cryptocurrency exchanges.

The primary target audience are researchers and students of Finance, the package compliments the likes of TTR, FinancialInstruments and PerformanceAnalytics.

The secondary target audience are data scientists that are looking to find high quality cryptocurrency data for developing and evaluating prediction models.

The third target audience are traders that are looking to develop and evaluate trading strategies using cryptocurrency data.

The packages which comes closest is the [coinmarketcapr](https://github.com/amrrs/coinmarketcapr)- and [crypto2](https://github.com/sstoeckl/crypto2)-packages.

However, `coinmarketcapr` is gated behind an API subscription and haven't been updated in the last two years. The `crypto2`-package, a fork of [crypto](https://github.com/JesseVent/crypto),  scrapes the data from [coinmarketcap](https://coinmarketcap.com/); the package saw its last update 4 months ago, and prior to that it have not been updated since March 2023. Both packages only supports, to the best of my knowledge, daily data.

The `cryptoQuotes`-package, albeit new as it was released December 2023, retrieves the data  by interacting with the public API endpoints provided by the cryptocurrency exchanges. Furthermore, the available data comes with a higher degree of quality and detail as the data is provided directly from the exchanges, from which  the data can be extracted in 1 second to 1 month intervals, which can be a necessity when studying price dynamics.

Furthermore, `cryptoQuotes` is not gated behind a subscription, and all its functions and features are extensively documented.

> **NOTE:** See issue https://github.com/sstoeckl/crypto2/issues/20 and https://github.com/cran/crypto2/issues/1. It seems that `crypto2` lost its access to the coinmarketcap API; which leaves `coinmarketcapr` as the only similar package left.

Firstly, `cryptoQuotes` has already been published on `CRAN`. ~~Secondly, I have an update pending (`1.3.1`) which is the `development`-branch of my repository. This update haven't been published as I wanted to submit and, hopefully review, my package with you before that update gets published!~~

**Edit:** I have published `1.3.1` to `CRAN` and moved the latest update to the `main`-branch.
