### 2024-05-26 80e9541
Deploying to gh-pages from @ serkor1/cryptoQuotes@6c8b62120e9004480148d860e56c1cf5ea458d6d 🚀
