### 2024-05-28 dc02d45
Updated chart coloring
The charts are now using hcl.colors which are robust to colorblindness. These will be implemented regardless of deficiency options.
