+++ R/chart_fgi.R
@@ -120,7 +120,7 @@ fgi <- function(
         annotations = list(
           text = "Fear and Greed Index",
           font = list(
-            size = 18
+            size = 16
           ),
           showarrow = FALSE,
           x = 0,
