+++ playground/playground.R
@@ -11,16 +11,25 @@ rm(list = ls()); gc(); devtools::load_all()
 
 
 chart(
-  ticker = BTC,
+  ticker = cryptoQuotes:::control_data$quote,
   main   = ohlc(),
   indicator = list(
+    bollinger_bands(),
     sma(n = 10),
-    sma(n = 15)
+    sma(n = 15),
+    sma(n = 14)
+  ),
+  sub = list(
+    macd(),
+    rsi(),
+    fgi(cryptoQuotes:::control_data$fgindex),
+    lsr(cryptoQuotes:::control_data$lsratio)
   ),
   options = list(
     deficiency = FALSE
   )
 )
+
 
 
 
