### 2024-05-28 e697291
Deploying to gh-pages from @ serkor1/cryptoQuotes@dc02d45af418ff991e69d23635c7a21ec6fe7ddb 🚀
