### 2024-05-28 0621448
Simplified lsr() and fgi() indicators
Both indicators were doing redundant calculations based on color deficiency. This have been address and is now more efficient.
