+++ R/chart_lsr.R
@@ -62,29 +62,17 @@ lsr <- function(
 
       # 1.1) define available
       # colors
-      color_scale <- grDevices::palette(
-        grDevices::hcl.colors(
-          n = 30,
-          palette = "RdYlGn"
-        )
+      color_scale <- grDevices::hcl.colors(
+        n = 30,
+        palette = ifelse(color_deficiency, "Cividis", "RdYlGn"),
+        rev = color_deficiency
       )
 
-      if (color_deficiency) {
-        color_scale <- rev(
-          grDevices::palette(
-            grDevices::hcl.colors(
-              n = 30,
-              palette = "Cividis"
-            )
-          )
+      data$color_scale <- normalize(
+        x = data$ls_ratio,
+        range = c(0,30),
+        value = c(0,3)
         )
-      }
-
-      # 1.1.1) map the values
-      # to  the color scale
-      data$color_scale <- ceiling(
-        30 * (data$ls_ratio)/(3)
-      )
 
       data$color_scale <- color_scale[
         data$color_scale
