### 2024-05-28 9587a8d
Deploying to gh-pages from @ serkor1/cryptoQuotes@0621448fd9d321b21f67dfa96e488b1c26eacead 🚀
