+++ pkgdown.yml
@@ -10,7 +10,7 @@ articles:
   cryptoQuotes: cryptoQuotes.html
   custom_indicators: custom_indicators.html
   usecase: usecase.html
-last_built: 2024-05-28T09:32Z
+last_built: 2024-05-28T10:37Z
 urls:
   reference: https://serkor1.github.io/cryptoQuotes/reference
   article: https://serkor1.github.io/cryptoQuotes/articles
