### 2024-05-29 c1d53ba
Streamlined indicators
All indicator functions now has a high level checker that asserts the presence of the chart-function.
