+++ playground/playground.R
@@ -11,28 +11,30 @@ rm(list = ls()); gc(); devtools::load_all()
 
 
 
-chart(
-  ticker = cryptoQuotes:::control_data$quote,
-  main   = ohlc(),
-  indicator = list(
-    bollinger_bands(),
-    sma(n = 10),
-    sma(n = 15),
-    sma(n = 14)
-  ),
-  sub = list(
-    macd(),
-    rsi(),
-    fgi(cryptoQuotes:::control_data$fgindex),
-    lsr(cryptoQuotes:::control_data$lsratio)
-  ),
-  options = list(
-    deficiency = FALSE
+bars <- function() {
+  chart(
+    ticker = cryptoQuotes:::control_data$quote,
+    main   = ohlc(),
+    indicator = list(
+      bollinger_bands(),
+      sma(n = 10),
+      sma(n = 15),
+      sma(n = 14)
+    ),
+    sub = list(
+      macd(),
+      rsi(),
+      fgi(cryptoQuotes:::control_data$fgindex),
+      lsr(cryptoQuotes:::control_data$lsratio)
+    ),
+    options = list(
+      deficiency = FALSE
+    )
   )
-)
 
+}
 
-
+bars()
 
 
 # script end;
