+++ R/chart_quote.R
@@ -20,24 +20,11 @@
 kline <- function(
     ...) {
 
-  call_stack <- as.character(
-    lapply(sys.calls(), `[[`, 1)
-  )
-
-  assert(
-    call_stack[1] != as.character(match.call()),
-    error_message = c(
-      "x" = "Error",
-      "i" = paste(
-        "Run",
-        cli::code_highlight(
-          code = "cryptoQuotes::chart(...)",
-          code_theme = "Chaos"
-        ),
-        "to build charts."
-      )
-    )
-  )
+  # check if the indicator is called
+  # from the chart-function
+  #
+  # stops the function if not
+  check_indicator_call()
 
   structure(
     .Data = {
@@ -137,24 +124,11 @@ kline <- function(
 ohlc <- function(
     ...) {
 
-  call_stack <- as.character(
-    lapply(sys.calls(), `[[`, 1)
-  )
-
-  assert(
-    call_stack[1] != as.character(match.call()),
-    error_message = c(
-      "x" = "Error",
-      "i" = paste(
-        "Run",
-        cli::code_highlight(
-          code = "cryptoQuotes::chart(...)",
-          code_theme = "Chaos"
-        ),
-        "to build charts."
-      )
-    )
-  )
+  # check if the indicator is called
+  # from the chart-function
+  #
+  # stops the function if not
+  check_indicator_call()
 
   structure(
     .Data = {
@@ -245,24 +219,11 @@ pline <- function(
     price = "close",
     ...) {
 
-  call_stack <- as.character(
-    lapply(sys.calls(), `[[`, 1)
-  )
-
-  assert(
-    call_stack[1] != as.character(match.call()),
-    error_message = c(
-      "x" = "Error",
-      "i" = paste(
-        "Run",
-        cli::code_highlight(
-          code = "cryptoQuotes::chart(...)",
-          code_theme = "Chaos"
-        ),
-        "to build charts."
-      )
-    )
-  )
+  # check if the indicator is called
+  # from the chart-function
+  #
+  # stops the function if not
+  check_indicator_call()
 
   structure(
     .Data = {
