### 2024-05-29 e7dea88
Deploying to gh-pages from @ serkor1/cryptoQuotes@c1d53ba87f28f85ffe286ef3176d1b0942b3155d 🚀
