### 2024-05-29 b078d5b
New Indicator
A new indicator have been introduced from the TTR-library. The Stochastic Momentum Index, the function is smi(). News and references have been updated.

Changes to functions: all functions are now steel blue by default, instead of purple.
