+++ playground/playground.R
@@ -9,32 +9,25 @@
 # setup;
 rm(list = ls()); gc(); devtools::load_all()
 
+ticker <- cryptoQuotes:::control_data$quote
 
-
-bars <- function() {
-  chart(
-    ticker = cryptoQuotes:::control_data$quote,
-    main   = ohlc(),
-    indicator = list(
-      bollinger_bands(),
-      sma(n = 10),
-      sma(n = 15),
-      sma(n = 14)
-    ),
-    sub = list(
-      macd(),
-      rsi(),
-      fgi(cryptoQuotes:::control_data$fgindex),
-      lsr(cryptoQuotes:::control_data$lsratio)
-    ),
-    options = list(
-      deficiency = FALSE
-    )
+chart(
+  ticker = ticker,
+  main   = kline(),
+  indicator = list(
+    bollinger_bands(),
+    sma(n = 10),
+    sma(n = 15),
+    sma(n = 14)
+  ),
+  sub = list(
+    smi(),
+    rsi()
+  ),
+  options = list(
+    deficiency = FALSE
   )
-
-}
-
-bars()
+)
 
 
 # script end;
