+++ man/fgi.Rd
@@ -100,6 +100,7 @@ Other chart indicators:
 \code{\link{macd}()},
 \code{\link{rsi}()},
 \code{\link{sma}()},
+\code{\link{smi}()},
 \code{\link{volume}()},
 \code{\link{vwap}()},
 \code{\link{wma}()},
@@ -113,6 +114,7 @@ Other subchart indicators:
 \code{\link{lsr}()},
 \code{\link{macd}()},
 \code{\link{rsi}()},
+\code{\link{smi}()},
 \code{\link{volume}()}
 }
 \author{
