### 2024-05-29 40c9a5e
Deploying to gh-pages from @ serkor1/cryptoQuotes@b078d5bf47b18b55ec6677c5eca53431565f5493 🚀
