### 2024-05-30 5d41fd0
Added Donchian Channels
A new mainchart-indicator has been introduced to library: Donchian Channels. See more here https://www.investopedia.com/terms/d/donchianchannels.asp
NEWS has been updated accordingly.

Minor additions:

Minor typos in NEWS.md has been fixed.
