+++ man/bollinger_bands.Rd
@@ -56,6 +56,7 @@ Other chart indicators:
 \code{\link{alma}()},
 \code{\link{chart}()},
 \code{\link{dema}()},
+\code{\link{donchian_channel}()},
 \code{\link{ema}()},
 \code{\link{evwma}()},
 \code{\link{fgi}()},
@@ -74,6 +75,7 @@ Other main chart indicators:
 \code{\link{add_event}()},
 \code{\link{alma}()},
 \code{\link{dema}()},
+\code{\link{donchian_channel}()},
 \code{\link{ema}()},
 \code{\link{evwma}()},
 \code{\link{hma}()},
