### 2024-05-30 e42677c
Deploying to gh-pages from @ serkor1/cryptoQuotes@da6e744ceb9e0bfc843496abd69c26775ef49180 🚀
