### 2024-05-30 bfc93e8
Deploying to gh-pages from @ serkor1/cryptoQuotes@5d41fd017c0dba6f27e04455166a73a1893f1c19 🚀
