### 2024-05-30 1e60df0
Updated Tests and FUN
The utility functions now have assert statements in the start. There were some bugs in the example of the utility functions. These have been fixed.

News not updated.
