+++ playground/playground.R
@@ -11,6 +11,42 @@ rm(list = ls()); gc(); devtools::load_all()
 
 ticker <- cryptoQuotes:::control_data$quote
 
+
+setequal(
+  window(
+    ticker,
+    start = "2024-02-09 01:00:00",
+    end   = "2024-02-16 01:00:00"
+  ),
+  ticker[paste(c("2024-02-09 01:00:00", "2024-02-16 01:00:00"), collapse = "/")]
+
+)
+
+
+foo <- function() {
+
+  window(
+    ticker,
+    start = "2024-02-09 01:00:00",
+    end   = "2024-02-16 01:00:00"
+  )
+
+}
+
+
+bar <- function() {
+
+  ticker[paste(c("2024-02-09 01:00:00", "2024-02-16 01:00:00"), collapse = "/")]
+
+}
+
+
+microbenchmark::microbenchmark(
+  foo(),
+  bar()
+)
+
+
 chart(
   ticker = ticker,
   main   = kline(),
