+++ man/calibrate_window.Rd
@@ -54,7 +54,7 @@ FGIndex <- cryptoQuotes::split_window(
 # it needs to passed into calibrateWindow
 # to ensure comparability
 FGIndex <- cryptoQuotes::calibrate_window(
-  list = cryptoQuotes::FGIndex,
+  list = FGIndex,
 
   # As each element in the list can include
   # more than one row, each element needs to be aggregated
@@ -68,10 +68,13 @@ FGIndex <- cryptoQuotes::calibrate_window(
 
 # 3) check if candles aligns
 # accordingly
-setequal(
-  zoo::index(BTC),
-  zoo::index(FGIndex)
+stopifnot(
+  setequal(
+    zoo::index(BTC),
+    zoo::index(FGIndex)
+  )
 )
+
 
 # script end;
 }
