+++ README.md
@@ -1,7 +1,7 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# cryptoQuotes: Open access to cryptocurrency market data <a href="https://serkor1.github.io/cryptoQuotes/"><img src="man/figures/logo.png" align="right" height="139" alt="cryptocurrency in R"/></a>
+# cryptoQuotes: Open access to cryptocurrency market data <a href="https://serkor1.github.io/cryptoQuotes/"><img src="man/figures/logo.png" align="right" height="154" alt="cryptocurrency in R"/></a>
 
 <!-- badges: start -->
 
@@ -279,122 +279,122 @@ volume
 <tbody>
 <tr>
 <td style="text-align:left;">
-2024-05-30 10:00:00
+2024-05-30 16:00:00
 </td>
 <td style="text-align:center;">
-67584.57
+68623.99
 </td>
 <td style="text-align:center;">
-67643.99
+68826.31
 </td>
 <td style="text-align:center;">
-67128
+68472.35
 </td>
 <td style="text-align:center;">
-67376.01
+68671.02
 </td>
 <td style="text-align:left;">
-1287.97865
+1274.59558
 </td>
 </tr>
 <tr>
 <td style="text-align:left;">
-2024-05-30 10:30:00
+2024-05-30 16:30:00
 </td>
 <td style="text-align:center;">
-67376.01
+68671.02
 </td>
 <td style="text-align:center;">
-67769.11
+68756.41
 </td>
 <td style="text-align:center;">
-67376
+68256.29
 </td>
 <td style="text-align:center;">
-67688.96
+68359.6
 </td>
 <td style="text-align:left;">
-644.27659
+758.51827
 </td>
 </tr>
 <tr>
 <td style="text-align:left;">
-2024-05-30 11:00:00
+2024-05-30 17:00:00
 </td>
 <td style="text-align:center;">
-67688.96
+68359.6
 </td>
 <td style="text-align:center;">
-67759.99
+68460.84
 </td>
 <td style="text-align:center;">
-67579.04
+68144
 </td>
 <td style="text-align:center;">
-67579.05
+68424
 </td>
 <td style="text-align:left;">
-414.02039
+638.09219
 </td>
 </tr>
 <tr>
 <td style="text-align:left;">
-2024-05-30 11:30:00
+2024-05-30 17:30:00
 </td>
 <td style="text-align:center;">
-67579.04
+68424.01
 </td>
 <td style="text-align:center;">
-67753.71
+68612
 </td>
 <td style="text-align:center;">
-67459.96
+68367.65
 </td>
 <td style="text-align:center;">
-67745.94
+68530.76
 </td>
 <td style="text-align:left;">
-468.4199
+911.26349
 </td>
 </tr>
 <tr>
 <td style="text-align:left;">
-2024-05-30 12:00:00
+2024-05-30 18:00:00
 </td>
 <td style="text-align:center;">
-67745.94
+68530.77
 </td>
 <td style="text-align:center;">
-67751
+68600
 </td>
 <td style="text-align:center;">
-67541.82
+68384
 </td>
 <td style="text-align:center;">
-67683.69
+68427.18
 </td>
 <td style="text-align:left;">
-314.39417
+411.48792
 </td>
 </tr>
 <tr>
 <td style="text-align:left;">
-2024-05-30 12:30:00
+2024-05-30 18:30:00
 </td>
 <td style="text-align:center;">
-67683.68
+68427.19
 </td>
 <td style="text-align:center;">
-67991
+68460
 </td>
 <td style="text-align:center;">
-67683.68
[…]
