### 2024-05-30 857ca3e
Deploying to gh-pages from @ serkor1/cryptoQuotes@1e60df0e3ed54389b790e1d5ab1ef33465fbd6e6 🚀
