### 2024-05-31 f9f9bca
Updated NEWS.md
The NEWS.md is now generated from NEWS.Rmd to accomodate changes and showcases
