+++ NEWS.md
@@ -1,63 +1,165 @@
+
 # cryptoQuotes 1.3.1
 
 ## General
 
-* Removed deprecated functions `getQuote()`-, `getFGIndex()`- and `getLSRatio()`-functions.
+> **NOTE:** With this update the package is no longer considered
+> `experimental`.
 
-* Removed decprecated functions `availableExchanges()`-, `availableIntervals()`- and `availableTickers()`-functions.
+- Removed deprecated functions `getQuote()`-, `getFGIndex()`- and
+  `getLSRatio()`-functions.
+
+- Removed decprecated functions `availableExchanges()`-,
+  `availableIntervals()`- and `availableTickers()`-functions.
+
+These functions were planned to be deleted in `1.4.0` - however, we have
+decided to aim for an `JOSS`- and `rOpenSci`-subscription which requires
+some degree of stability. Hence the deletion of these function at
+`1.3.0`.
+
+We expect `1.4.0` to be released *after* `rOpenSci`-submission and
+acceptance :pray:
 
 ## Improvements
 
 ### New features
 
-* `smi()`-function, a `subchart`-indicator built on the `TTR::SMI()`-function.
-* `donchian_channel()`-function, a `main chart`-indicator built on the `TTR::DonchianChannel()`-function
+- `smi()`-function, a `subchart`-indicator built on the
+  `TTR::SMI()`-function.
+- `donchian_channel()`-function, a `main chart`-indicator built on the
+  `TTR::DonchianChannel()`-function
+
+<details>
+<summary>
+Usage
+</summary>
+
+``` r
+chart(
+  ticker = BTC,
+  main   = kline(),
+  indicator = list(
+    donchian_channel()
+  ),
+  sub = list(
+    smi()
+  )
+)
+```
+
+<img src="man/figures/NEWS-unnamed-chunk-2-1.png" style="display: block; margin: auto;" />
+</details>
 
 ### Expanded Support
 
-* `get_openinterest()` is now supported by `kraken`
+- `get_openinterest()` is now supported by `kraken`
+
+<details>
+<summary>
+Usage
+</summary>
+
+``` r
+tail(
+  get_openinterest(
+    ticker   = "PF_XBTUSD",
+    interval = "1h",
+    source   = "kraken"
+  )
+)
+```
+
+    #>                     open_interest
+    #> 2024-05-31 14:00:00      3018.442
+    #> 2024-05-31 15:00:00      3013.342
+    #> 2024-05-31 16:00:00      2957.343
+    #> 2024-05-31 17:00:00      2960.819
+    #> 2024-05-31 18:00:00      2954.668
+    #> 2024-05-31 19:00:00      2981.381
+
+</details>
 
 ### Charting
 
-* The `chart()`-function now has proper `light`-theme available that isn't the default `plotly`-values.
-* The charts now displays date ranges as a subtitles.
-* The `bollinger_bands()`-function now accepts a `color`-argument. These can be passed as `Hexadecimal`-colors or as is, `"blue"` for example.
-* A new main chart function has been introduced. `pline()` which is a univariate price chart based on either open, high, low or close prices.
+- The `chart()`-function now has proper `light`-theme available that
+  isn’t the default `plotly`-values.
+- The charts now displays date ranges as a subtitles.
+- The `bollinger_bands()`-function now accepts a `color`-argument. These
+  can be passed as `Hexadecimal`-colors or as is, `"blue"` for example.
+- A new main chart function has been introduced. `pline()` which is a
+  univariate price chart based on either open, high, low or close
+  prices.
+
+<details>
+<summary>
+Usage
+</summary>
+
+``` r
+chart(
+  ticker = BTC,
[…]
