+++ NEWS.Rmd
@@ -1,0 +1,335 @@
+---
+output: github_document
+always_allow_html: true
+---
+
+```{r setup, include = FALSE}
+# set options
+Sys.setenv(OPENSSL_CONF="/dev/null")
+
+knitr::opts_chunk$set(
+  collapse  = FALSE,
+  comment   = "#>",
+  fig.path  = "man/figures/NEWS-",
+  message   = FALSE,
+  warning   = FALSE,
+  dpi       = 120,
+  fig.align = "center"
+)
+```
+
+```{r, include=FALSE, echo=FALSE}
+library(cryptoQuotes)
+```
+
+# cryptoQuotes 1.3.1
+
+## General
+
+> **NOTE:** With this update the package is no longer considered `experimental`.
+
+* Removed deprecated functions `getQuote()`-, `getFGIndex()`- and `getLSRatio()`-functions.
+
+* Removed decprecated functions `availableExchanges()`-, `availableIntervals()`- and `availableTickers()`-functions.
+
+These functions were planned to be deleted in `1.4.0` - however, we have decided to aim for an `JOSS`- and `rOpenSci`-subscription which requires some degree of stability. Hence the deletion of these function at `1.3.0`.
+
+We expect `1.4.0` to be released *after* `rOpenSci`-submission and acceptance :pray:
+
+## Improvements
+
+### New features
+
+* `smi()`-function, a `subchart`-indicator built on the `TTR::SMI()`-function.
+* `donchian_channel()`-function, a `main chart`-indicator built on the `TTR::DonchianChannel()`-function
+
+<details>
+<summary>Usage</summary>
+```{r}
+chart(
+  ticker = BTC,
+  main   = kline(),
+  indicator = list(
+    donchian_channel()
+  ),
+  sub = list(
+    smi()
+  )
+)
+```
+</details>
+
+### Expanded Support
+
+* `get_openinterest()` is now supported by `kraken`
+
+<details>
+<summary>Usage</summary>
+```{r}
+tail(
+  get_openinterest(
+    ticker   = "PF_XBTUSD",
+    interval = "1h",
+    source   = "kraken"
+  )
+)
+```
+</details>
+
+### Charting
+
+* The `chart()`-function now has proper `light`-theme available that isn't the default `plotly`-values.
+* The charts now displays date ranges as a subtitles.
+* The `bollinger_bands()`-function now accepts a `color`-argument. These can be passed as `Hexadecimal`-colors or as is, `"blue"` for example.
+* A new main chart function has been introduced. `pline()` which is a univariate price chart based on either open, high, low or close prices.
+
+<details>
+<summary>Usage</summary>
+```{r}
+chart(
+  ticker = BTC,
+  main   = pline(price = "close"),
+  indicator = list(
+    bollinger_bands(
+    color = "steelblue"
+  )
+  ),
+  sub = list(
+    volume()
+  ),
+  options = list(
+    dark = FALSE
+  )
+)
+```
+</details>
+
+### Documentation
+
+* The documentation has been extensively reworked. This is includes, but not limited to, sample outputs for all `get_*`-functions.
+
+### Backend Changes
+
+* The `chart()`-functions no longer depend on `rlang`. 
+
+> **Note:** we are generally moving away from `rlang`, `purrr` and `tidyverse` in general. We are, however, 
+> keeping the `tidyverse` styleguide.
+
+* Removed dependency on `conflicted`-package.
+
+Prior to version `1.3.0` the `get*`-functions were following the syntax of `quantmod` closely, and this goes for the function naming too. With the adoption of the `tidyverse` style guide, there is no conflicts that
+needs to be resolved on `stable`- and `experimental`-functions.
+
+### New developper tools
+
+* `pull()`-function
+* `var_ly()`-function
+* `build()`-function
+
[…]
