### 2024-05-31 cc414bf
Deploying to gh-pages from @ serkor1/cryptoQuotes@f9f9bcaabc82e342e0f6cee986cb838f44dc4929 🚀
