### 2024-05-31 9c8ac7b
CRAN Submission
This commit is the submitted version to CRAN. See https://github.com/serkor1/cryptoQuotes/issues/11 for a full list of tasks.

Please refer to the NEWS.md for a full list of changes.
