+++ R/chart.R
@@ -31,11 +31,7 @@
 #'
 #' **Sample Output**
 #' \if{html}{
-#'   \out{<div style="text-align: center">}
-#'    \figure{README-chartquote-1.png}{
-#'    options: style="width:750px;max-width:75\%;"
-#'    }
-#'   \out{</div>}
+#'   \out{<span style="text-align: center; display: block;">}\figure{README-chartquote-1.png}{options: style="width:750px;max-width:75\%;"}\out{</span>}
 #' }
 #' \if{latex}{
 #'   \out{\begin{center}}\figure{README-chartquote-1.png}\out{\end{center}}
