### 2024-05-31 970bfb8
CRAN Submission
This commit is the submitted version to CRAN. See https://github.com/serkor1/cryptoQuotes/issues/11 for a full list of tasks.

Please refer to the NEWS.md for a full list of changes.
