+++ README.md
@@ -10,7 +10,7 @@ status](https://www.r-pkg.org/badges/version/cryptoQuotes)](https://CRAN.R-proje
 [![CRAN RStudio mirror
 downloads](https://cranlogs.r-pkg.org/badges/last-month/cryptoQuotes?color=blue)](https://r-pkg.org/pkg/cryptoQuotes)
 [![R-CMD-check](https://github.com/serkor1/cryptoQuotes/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/serkor1/cryptoQuotes/actions/workflows/R-CMD-check.yaml)
-[![codecov](https://codecov.io/gh/serkor1/cryptoQuotes/graph/badge.svg?token=D7NF1BPVL5)](https://codecov.io/gh/serkor1/cryptoQuotes)
+[![codecov](https://codecov.io/gh/serkor1/cryptoQuotes/graph/badge.svg?token=D7NF1BPVL5)](https://app.codecov.io/gh/serkor1/cryptoQuotes)
 [![Lifecycle:
 stable](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable)
 ![GitHub last commit
@@ -279,122 +279,122 @@ volume
 <tbody>
 <tr>
 <td style="text-align:left;">
-2024-05-30 16:00:00
+2024-05-31 18:00:00
 </td>
 <td style="text-align:center;">
-68623.99
+67215.58
 </td>
 <td style="text-align:center;">
-68826.31
+67352.41
 </td>
 <td style="text-align:center;">
-68472.35
+66670
 </td>
 <td style="text-align:center;">
-68671.02
+67000.01
 </td>
 <td style="text-align:left;">
-1274.59558
+2093.77494
 </td>
 </tr>
 <tr>
 <td style="text-align:left;">
-2024-05-30 16:30:00
+2024-05-31 18:30:00
 </td>
 <td style="text-align:center;">
-68671.02
+67000.01
 </td>
 <td style="text-align:center;">
-68756.41
+67221.53
 </td>
 <td style="text-align:center;">
-68256.29
+66876.8
 </td>
 <td style="text-align:center;">
-68359.6
+67200.01
 </td>
 <td style="text-align:left;">
-758.51827
+788.25693
 </td>
 </tr>
 <tr>
 <td style="text-align:left;">
-2024-05-30 17:00:00
+2024-05-31 19:00:00
 </td>
 <td style="text-align:center;">
-68359.6
+67200.01
 </td>
 <td style="text-align:center;">
-68460.84
+67459.66
 </td>
 <td style="text-align:center;">
-68144
+67160.3
 </td>
 <td style="text-align:center;">
-68424
+67417.98
 </td>
 <td style="text-align:left;">
-638.09219
+598.68095
 </td>
 </tr>
 <tr>
 <td style="text-align:left;">
-2024-05-30 17:30:00
+2024-05-31 19:30:00
 </td>
 <td style="text-align:center;">
-68424.01
+67417.98
 </td>
 <td style="text-align:center;">
-68612
+67455.93
 </td>
 <td style="text-align:center;">
-68367.65
+67287.97
 </td>
 <td style="text-align:center;">
-68530.76
+67342.77
 </td>
 <td style="text-align:left;">
-911.26349
+287.26257
 </td>
 </tr>
 <tr>
 <td style="text-align:left;">
-2024-05-30 18:00:00
+2024-05-31 20:00:00
 </td>
 <td style="text-align:center;">
-68530.77
+67342.77
 </td>
 <td style="text-align:center;">
-68600
[…]
