+++ NEWS.Rmd
@@ -2,6 +2,7 @@
 output: github_document
 always_allow_html: true
 ---
+<!-- NEWS.md is generated from NEWS.Rmd. Please edit that file -->
 
 ```{r setup, include = FALSE}
 # set options
@@ -14,7 +15,7 @@ knitr::opts_chunk$set(
   message   = FALSE,
   warning   = FALSE,
   dpi       = 120,
-  fig.align = "center"
+  fig.align = 'center'
 )
 ```
 
