+++ NEWS.md
@@ -1,4 +1,6 @@
 
+<!-- NEWS.md is generated from NEWS.Rmd. Please edit that file -->
+
 # cryptoQuotes 1.3.1
 
 ## General
@@ -70,12 +72,12 @@ tail(
 ```
 
     #>                     open_interest
-    #> 2024-05-31 14:00:00      3018.442
     #> 2024-05-31 15:00:00      3013.342
     #> 2024-05-31 16:00:00      2957.343
     #> 2024-05-31 17:00:00      2960.819
     #> 2024-05-31 18:00:00      2954.668
-    #> 2024-05-31 19:00:00      2981.381
+    #> 2024-05-31 19:00:00      2983.686
+    #> 2024-05-31 20:00:00      2996.449
 
 </details>
 
