### 2024-05-31 35d0dd2
Deploying to gh-pages from @ serkor1/cryptoQuotes@970bfb8ce940ad8f6785e29fdfb726868c8871c9 🚀
