### 2024-05-31 22ba9b3
Deploying to gh-pages from @ serkor1/cryptoQuotes@c4ab323b5f4e8b870b5f53fef6cd5a8a0a5cdfed 🚀
