### 2024-05-31 c326df5
Deploying to gh-pages from @ serkor1/cryptoQuotes@9c8ac7b292671ea00cb498f0d678f93b7a9dfa24 🚀
