### 2024-05-31 7af5c47
Update playground.R
Solves PR https://github.com/serkor1/cryptoQuotes/pull/12
