### 2024-05-31 8222fbe
Deploying to gh-pages from @ serkor1/cryptoQuotes@7af5c4747557b690749f5db3469a0078ca4485e7 🚀
