### 2024-05-31 93d7baa
Version `1.3.1` (#12)
See https://github.com/serkor1/cryptoQuotes/blob/main/NEWS.md for more details.
