+++ R/chart_lsr.R
@@ -3,115 +3,119 @@
 # author: Serkan Korkmaz, serkor1@duck.com
 # objective:
 # script start;
-
-
-#' Chart the long-short ratios
+#' @title
+#' Chart the long-short ratio
 #'
 #' @description
-#'
 #' `r lifecycle::badge("experimental")`
 #'
-#' The [lsr()]-function adds a scatter plot as a subplot to the chart colored by ratio size.
+#' A high-level [plotly::plot_ly()]-wrapper function.
+#' The function adds a subchart to the [chart] with `long-short ratio`.
 #'
-#' @details
-#' The long-short ratio is a market sentiment indicator on expected price movement.
+#' @param ratio A [xts::xts()]-object. See [get_lsratio()] for more details.
+#' @inherit kline
 #'
-#' @param ratio A [xts::xts()]-object with the column LSRatio. See [get_lsratio()] for more details.
-#' @param internal An empty [list()]. This is an internal helper-argument, ignore.
-#'
-#' @example man/examples/scr_LSR.R
+#' @example man/examples/scr_chartLSR.R
 #'
 #' @family chart indicators
 #' @family sentiment indicators
-#' @family subcharts
-#'
-#' @returns A [plotly::plot_ly()]-object wrapped in [rlang::expr()]
-#'
+#' @family subchart indicators
+#' @author Serkan Korkmaz
 #' @export
 lsr <- function(
     ratio,
-    internal = list()) {
+    ...) {
+
+  # check if the indicator is called
+  # from the chart-function
+  #
+  # stops the function if not
+  check_indicator_call()
+
   structure(
-    rlang::expr(
-      {
+    .Data = {
 
-        # 0) locate global
-        # parameters to be passed
-        # into the charting functions;
+      # 0) construct arguments
+      # via chart function
+      args <- list(
+        ...
+      )
 
-        # 0) locate global
-        # parameters to be passed
-        # into the charting functions;
-
-        # internal_args <- flatten(lapply(!!rlang::enquos(internal), rlang::eval_tidy))
-        internal_args <- flatten(lapply(!!rlang::enexpr(internal), rlang::eval_tidy))
-
-        ticker <- internal_args$ticker
-        deficiency <- internal_args$deficiency
+      # 1) unpack values
+      # from chart
+      data <- zoo::fortify.zoo(ratio, names = "index")
+      color_deficiency <- args$deficiency
 
 
-        ratio <- toDF(
-          quote = !!ratio
+      # 1.1) define available
+      # colors
+      color_scale <- grDevices::hcl.colors(
+        n = 30,
+        palette = ifelse(color_deficiency, "Cividis", "RdYlGn"),
+        rev = color_deficiency
+      )
+
+      data$color_scale <- normalize(
+        x = data$ls_ratio,
+        range = c(0,30),
+        value = c(0,3)
         )
 
+      data$color_scale <- color_scale[
+        data$color_scale
+      ]
 
-        value <- ratio$ls_ratio
-
-        ratio$color_scale <- round((30) * (value -0)/(2 - 0))
-
-
-        color_scale <- if (deficiency){
-
-
-          rev(paletteer::paletteer_c("grDevices::Cividis", 30))
-
-
-        } else {
-          paletteer::paletteer_c("grDevices::RdYlGn", 30)
-
-        }
-
-
-
-
-        ratio$color_scale <- color_scale[
-          ratio$color_scale
-        ]
-
-
-
-        plotly::layout(
-          yaxis = list(
-            title = 'Long-Short Ratio'
+      plotly::layout(
+        plotly::plot_ly(
+          showlegend = FALSE,
+          name = "Long-Short Ratio",
+          data = data,
+          y = ~ls_ratio,
[…]
