### 2024-06-02 #644 @serkor1
Hi @jooolia,

I see that `phangorn`-package got its presubmission approval the same day it was labelled (Issue https://github.com/ropensci/software-review/issues/641). Have I missed something?

Best,
