### 2024-06-07 d7099e1
Deploying to gh-pages from @ serkor1/cryptoQuotes@93d7baa82ca0811370a19a3a4cd5f65d6e7a908f 🚀
