### 2024-06-07 037d7db
Expanded Exchange Support [See https://github.com/serkor1/cryptoQuotes/issues/14]
The following exchanges have been addded

* crypto.com (spot, futures, fundingrate)
* huobi (spot, futures)
* MEXC (spot, futures, fundingrate)

Minor changes

get_quote()-function needed small changes in order to return the data in OHLC-V format
due to Huobi API specifics.

NOTE:
crypto.com US compliant and can therefore run on US resident computers without VPN.
Huobi API documentation is huge - this still needs work, and confirmation on what it can return
other than what is provided already. The funding rate fo
[…]
