+++ playground/playground.R
@@ -9,45 +9,36 @@
 # setup;
 rm(list = ls()); gc(); devtools::load_all()
 
-BTC <- cryptoQuotes::get_quote(
-  ticker   = 'BTCUSD',
-  interval = '1d',
-  source   = 'kraken',
-  futures  = FALSE,
-  from     = Sys.Date() - 1000,
-  to       = Sys.Date()
+
+get_fundingrate(
+  "BTC_USDT",
+  "mexc"
 )
 
-nrow(BTC)
-
-
-head(BTC)
-head(BTC)
-
-
-
-length(
-  seq(
-    from = coerce_date(Sys.Date() - 10),
-    to   = coerce_date(Sys.Date() - 1),
-    by   = "+15 mins"
-  )
+get_quote(
+  "BTC_USDT",
+  source = "mexc",
+  TRUE
 )
 
-
-default_dates(
-  interval = "15m",
-  from = coerce_date(Sys.Date() - 10),
-  to   = coerce_date(Sys.Date() - 1)
+get_quote(
+  "BTCUSDT",
+  source = "mexc",
+  FALSE
 )
+
+get_quote(
+  ticker = "BTC-USDT",
+  source = "huobi",
+  futures = TRUE,
+  interval = "1h"
+)
+
+get_quote(
+  ticker = "btcusdt",
+  source = "huobi",
+  futures = FALSE
+)
+
 
 # script end;
-
-
-origin_date <- '1970-01-01'
-
-as.POSIXct(
-  trunc(as.double(coerce_date(Sys.Date() - 10))/(15*60))*(15*60),
-  tz = 'UTC',
-  origin = origin_date
-)
