+++ R/api_huobi.R
@@ -1,0 +1,247 @@
+# script: api_huobi
+# date: 2024-06-07
+# author: Serkan Korkmaz, serkor1@duck.com
+# objective: Create all necessary parameters
+# for a proper API call
+# script start;
+# 1) URLs and Endpoint; ####
+huobiUrl <- function(
+    futures = TRUE,
+    ...) {
+
+  # 1) define baseURL
+  # for each API
+  base::ifelse(
+    test = futures,
+    yes  = 'https://api.hbdm.com',
+    no   = 'https://api.huobi.pro'
+  )
+
+}
+
+huobiEndpoint <- function(
+    type    = 'ohlc',
+    futures = TRUE,
+    top     = FALSE) {
+
+ switch(
+    EXPR = type,
+    ohlc = {
+      if (futures)
+        'linear-swap-ex/market/history/kline'
+      else
+        'market/history/kline'
+    },
+    ticker ={
+      if (futures)
+        "linear-swap-api/v1/swap_api_state"
+      else
+        "v2/settings/common/symbols"
+    }
+  )
+}
+
+# 2) Available intervals; #####
+huobiIntervals <- function(
+    futures,
+    interval,
+    all = FALSE,
+    type,
+    ...) {
+
+  # 0) define intervals
+  # NOTE: These are common for
+  # both endpoints
+  all_intervals <- data.frame(
+    labels = c(
+      '1m',
+      '5m',
+      '15m',
+      '30m',
+      '1h',
+      '4h',
+      '1d',
+      '1w',
+      '1M'
+    ),
+    values =  c(
+      "1min",
+      "5min",
+      "15min",
+      "30min",
+      "60min",
+      "4hour",
+      "1day",
+      "1week",
+      "1mon"
+    )
+  )
+
+
+  if (all) {
+
+    all_intervals$labels
+
+  } else {
+
+    # Select the specified interval
+    selectedInterval <- all_intervals$values[
+      grepl(paste0('^', interval, '$'), all_intervals$labels)
+    ]
+
+    selectedInterval
+  }
+}
+
+# 3) define response object and format; ####
+huobiResponse <- function(
+    type = 'ohlc',
+    futures,
+    ...) {
+
+  # mock response
+  # to avoid check error in
+  # unevaluated expressions
+  response <- NULL
+
+  switch(
+    EXPR = type,
+    ohlc = {
+      list(
+        colum_names = c(
+          'open',
+          'close',
+          'low',
+          'high',
+          'volume'
+        ),
+        colum_location = if (futures) c(2:5,7) else c(2:6),
+        index_location = c(
+          1
+        )
+
+      )
+    },
+
+    ticker = {
+      list(
+        foo = function(
+    response,
+    futures = NULL){
+
+          if (futures) {
+
+            subset(
+              response$data
+            )$contract_code
+
+          } else {
+
+            subset(
+              response$data,
+              response$data$state == "online" & response$data$te == TRUE
+            )$sc
+
+          }
+
+        }
+      )
+    }
+  )
+
+}
+
+# 4) Dates passed to and from endpoints; ####
+huobiDates <- function(
+    futures,
+    dates,
+    is_response = FALSE,
+    ...) {
+
+  # 0) Set multiplier
+  multiplier <- 1
+
+  # 1) determine wether
+  # its a response or request
+  if (is_response) {
+
+    # NOTE: The API returns
+    # in GMT+8 time and
+    # convert_date assumes that its
+    # UTC
+    dates <- convert_date(
+      x = as.numeric(dates),
+      multiplier = multiplier)
+
+  } else {
+
+    dates <- convert_date(
+      x = dates,
+      multiplier = multiplier)
+
+    dates <- vapply(
+      dates,
[…]

[…]
