### 2024-06-11 #644 @sstoeckl
Dear @serkor1 and @jooolia, I have just pushed fixes for the broken @crypto2 package and, once again, all available historical prices, listing information etc is available for all cryptocurrencies (this is a truly survivorship bias-free dataset for research purposes). It is used to create asset-pricing factors for cryptos etc. If interesting I will consider submitting it to ropensci once I have it back on CRAN.
